To conduct a free and fair elections we have created a WebApp, this allows its users to report these activities with ease.  Election Commission, a neutral entity that conducts these elections will take note of the reported findings and act upon them.
To conduct a free and fair elections we have created a WebApp, this allows its users to report these activities with ease.
Election Commission, a neutral entity that conducts these elections will take note of the reported findings and act upon them. Through this app people can report instances and it will automatically be added to the database which is kept centralised to the government. Your location and live hotspots are detected through Google Maps API, it tells the precise location of the constituency/area of the event
To check the authenticity of the user a two step verification process is used which makes it as safe as a bank transaction, an OTP is sent to their respected e-mail ID and after verifying the OTP the complaint is registered and taken notice of
There is a live news feed available to keep people updated regarding the elections and other national news.
Statistics of the reported activities are also available, this is made by canvas.js; the statistics given are of those events that are reported by the people. The statistics are shown in a form of a colourful pie chart which zooms in when hovered